[usdMaya] Allow writing to already-open layers
Previously, the usdExport command errored when writing to a layer that was already open,
since we couldn't create a new layer with that identifier.

Now we simply take the existing layer and Clear() it (which guarantees that it is now
identical to a layer that was CreateNew()'ed) and then write into it. This gives similar
semantics to the append=True option, which writes into the open layer if it is in memory.

Once the layer is written, any stages that reference it should automatically
recompose. This seems like it should reduce potential confusion that might be caused
by having a different version of the layer on disk compared to in-memory. As a bonus,
you can now usdExport in regular "overwrite" mode to an anonymous layer.
(Previously, you could only append to anonymous layers.)

Fixes #636

(Internal change: 1912612)