[usdMaya/pxrUsdTranslators] - refactor prim writers to operate on MFnDependencyNodes
This change refactors the prim writers and prim writer registry system to
work in terms of MFnDependencyNodes (and instances of derived classes)
instead of in terms of MDagPaths. This is more work in the direction of
being able to support type-based prim writer lookup for Maya DG nodes in
the same way that we currently support writers for DAG nodes.

Note that for DAG nodes, an MDagPath is required to uniquely identify
instances in the DAG, so DAG node writers should be created using an
MFnDagNode (or one of its derived classes) constructed using an MDagPath,
*not* an MObject. There is some logic in the base UsdMayaPrimWriter class
as well as in the registry to verify that MFnDagNodes have valid MDagPaths
and to issue errors when that's not the case.

(Internal change: 1913040)