[Hd/Hdx] Removed children for Hydra Task.
Prior to this change, Hydra tasks were designed to be hierarchical.  Where a task could have a list of child tasks.

This was never really taken advantage of and planned optimizations to the Sync algorithm benefit from having a flat task namespace.
While it doesn't preclude delegates from using a task hierarchy, it does put the burden on them to flatten them before passing them to Hydra.

Changes:
- Removed _children from HdTask.  Sync and Execute no longer forward these calls to other tasks.
- HdTask no longer gets the "children" attribute from the scene delegate.
- Removed children dirty bit from change tracker.
- Remove children attribute and dirty bit handling from all uses of HdTask.

(Internal change: 1913113)