Properly handle usd edits to material binding changes
UsdImaging currently does a deferred population of materials. Materials are not populated unless a gprim is bound to them, also this population only happens when the gprim is being populated.

This change forces a prim resync when there is a change in a material binding, this guarantees correctness at a price. Long term we are looking at alternatives to the deferred population.

Finally, this changelist adds a new test called testUsdviewMaterialEdits which edits material bindings.

(Internal change: 1913118)