[usdMaya] plumb the UsdMayaWriteJobContext through to shading mode exporters
This change makes the UsdMaya_WriteJob pass through its entire write job
context through UsdMayaTranslatorMaterial and down to the shading mode
exporter. This context contains both the USD stage and the shadingMode, as
well as all of the export arguments, so it's no longer necessary to have
the separate UsdMayaExportParams class used by
UsdMayaShadingModeExporterContext. Any documentation for members of
UsdMayaExportParams was moved to the corresponding member of
UsdMayaJobExportArgs.

Giving shading mode exporters access to the full UsdMayaWriteJobContext
will make it possible for them to access the prim writer registry and
create prim writers for shaders based on their type name.

(Internal change: 1913165)