[Hd/UsdImaging] Optimization for Batch Removal of prims
When remove prims from Hydra's Render Index, it is often a subtree that gets removed.  Currently UsdImaging flattens that subtree and removes prims from Hydra one at a time.  Part of this removal process is removing the Prims Id for the Hd_SortedId list.

A recent change to SdfPath revealed how vulnerable the Hd_SortedIds was to path comparison performance under the batch removal scenarios.  The first removal would be a binary search, but all subsequent removals would perform a linear search.

This optimization takes advantage of the typical pattern of a batch removal, when the next prim to be removed, is typically the one following the one that just been removed.  When Hd_SortedIds performs a remove, it stores the location the location of the id after the id it just removed.  It then uses this as the first a hint point for the next remove.

In addition, Hd_SortedIds is already tracking how much of the list is sorted vs unsorted.  Therefore, if the id to remove wasn't the next item, but it is in the sorted portion of the list, then a binary search is done on the sorted portion of the list.  Only resulting to the linear search of the unsorted portion if the path is in the sorteded section of the list.

Changes:
- Hd_SortedIds now has a _afterLastDeletePoint member.
- Operations that invalidate the list, reset _afterLastDeletePoint.
- Updated Hd_SortedIds::Remove() to take advantage of _afterLastDeletePoint and binary searching sorted portion.
- Added additional performance tracking scopes to UsdImagingIndexProxy::_ProcessRemovals()
- Updated SortedId's test to include a remove after insert test.

(Internal change: 1913192)