[usdMaya] add utility methods for authoring USD shading networks
This change adds two convenience methods for authoring UsdShadeMaterial and
UsdShadeShader inputs and outputs and creating connections between them.
Material prims are meant to act as interfaces to the shading network, so
inputs on the Material prim will typically be the source for connections to
inputs on Shader prims inside the network.
CreateMaterialInputAndConnectShader() can be used to author that for a
given Material and Shader prim pair. Similarly, Shader prim outputs will
typically be the source for connections to outputs on Material prims.
CreateShaderOutputAndConnectMaterial() can be used to author that for a
given Shader and material prim pair.

(Internal change: 1913194)