[Hd/Hdx/UsdImaging + Tests] Move Task id down from HdSceneTask to HdTask.
One of the goals of this current series of checkins is to standardize the Sync API for tasks and make
it align more with other prims in sync.

With this goal HdSceneTask, is redundant, so these moves are in place to remove it.  Though, there is
some entanglement that needs sorting.

This change move the task Id from HdSceneTask to HdTask, so that the render index's SyncAll function may access it, without the
need of casting.

For tasks that are not a part of the scene graph, any id could be used (as long as it doesn't clashes with one in the scene), but the
empty path is recommended.  As a result, this change modifies tests and other direct users of HdTask to use the empty path.

Changes:
- Made HdTask() default constructor deleted.
- HdTask constructor now takes an id.
- HdSceneTask now forwards the id it gets in the constructor to the base class.
- Moved _id attribute and its getter from HdSceneTask to HdTask.
- There was 1 direct usage of _id in HdSceneTask in _MarkClean and that's been changed to use the getter.
- Hdx/Intersector and _DebugGroupTaskWrapper(in UsdImaging/engine) set the empty path for their tasks.
- Lots of tests use HdTask directly and now set the empty path

(Internal change: 1914075)