[Hd] Move change tracker clearing into Render Index Sync.
One of the goals of this current series of checkins is to standardize the Sync API for tasks and make
it align more with other prims in sync.

With this goal HdSceneTask, is redundant, so these moves are in place to remove it.  Though, there is
some entanglement that needs sorting.

In general when Hydra sync's prims, it will provide the Sync algorithm a modifiable set of dirty bits and once completed the sync algorithm will update the change tracker on behalf of the prim.

For tasks, that work is currently being done by HdSceneTask and always sets it to all clean, so this change moves it into the Sync algorithm.

The prior change in the series moved task id into HdTask, so now it is possible for SyncAll to lookup a prim in the render index and clear the change tracker.

This change doesn't change the Task's Sync API to pass in the dirty bits or do partial cleaning, that is left for another checkin.

Changes:
- Removed _MarkClean() member function from HdTask and HdSceneTask.
- HdTask::Sync() no longer calls _MarkClean() and is now just a forwarding function to the _Sync() virtual (to be addressed later).
- HdRenderIndex::SyncAll() now looks up the task in the render index based off its Id.  If the task is present in the render index, it calls that instance of the task.  Once the task is Sync'ed it is cleaned.
- For tasks, not found in the render index, Sync is called directly as before.

Notes:
A subtle change in behavior with this change is that previously hydra didn't use the tasks in the render index.  It only processed the passed in tasks.  So it was possible to have to tasks use the same path (/a), but pass different pointers for /a each time.  Now, the same instance (pointer) for /a (the one created when the task is inserted) is always used.

(Internal change: 1914118)