Remove use of raw getenv in favour of TfEnvSetting
The PXR_NDR_SKIP_DISCOVERY_PLUGIN_DISCOVERY environment variable
(which is only mainly used for testing) is now a TfEnvSetting instead
of having its value be determined through a raw call to getenv.  This
gives us an opportunity to add documentation for the env var as well
as make sure that its values are treated consistently, and finally, allows
for platform specific tweaks to how we deal with env vars apply to this
env var as well.

(Internal change: 1914292)