[usdMaya] introduce "useRegistry" shading mode exporter
This change adds a new "useRegistry" shading mode exporter that uses
UsdMayaShaderWriters registered in the prim writer registry to export
Maya shading dependency graphs. Similar to how the type-based DAG node
export works during a usdExport, this new shading exporter performs a DG
traversal for each shadingEngine node and looks for shader writers for the
shading nodes it encounters based on their Maya type names.

UsdMayaShaderWriter was added as a specialization of UsdMayaPrimWriter that
helps provide the translation between Maya attribute names and their
corresponding properties in USD. This allows getting the USD properties for
the source and destination of a shading connection in Maya and authoring a
corresponding connection in USD.

Note that this change adds export support only. Import support will be
added in a subsequent change.

(Internal change: 1914312)