[usdMaya] remove the error message when _getMayaShadersColor() heuristics fail
The _getMayaShadersColor() helper function that is used by
UsdMayaUtil::GetLinearShaderColor() assumes that the shaders it's given
either support the MFnLambertShader function set, or have attributes named
"color" and "transparency". Since that is not the case for all possible
shaders, it doesn't make sense to emit an error message when those
assumptions turn out to be incorrect. Instead, the function now just
returns whether it was able to read *any* color values successfully or not.

(Internal change: 1914314)