[Hd/Hdx] Add scene delegate and dirty bits to Task's Sync API.
Continuing the goal of bringing the Task system more inline with the rest of the Sync algorithm,  this change adds the passing of the
scene delegate and task dirty bits to the Task through the Sync method.

In line with other sync methods, it is now the Task's responsibility to clear the dirty bits and as such, task now support partial updates.

With this change, tasks currently have two ways to get their delegate and dirty bits.  From those passed to Sync and obtained from HdSceneTask.  Future change lists will migrate these tasks on to using the API versions.

This change necessitated the removal of _DebugGroupTaskWrapper as this concept causes problems for task delegate management and dirty bit tracking

Changes:
- HdTask::Sync now also takes a HdSceneDelegate * and HdDirtyBits * parameter.
- HdRenderIndex::SyncAll() now gets a tasks dirty bit from the change tracker and updates the change tracker with the returned value.
- For untracked task, the delegate is set to nullptr and a dummy dirtybits is used (set to 0).
- All tasks have been updated for the new parameters.
- Tasks that created a local variable called 'delegate' that they just obtained from HdSceneTask have been updated to use the new parameter.
- Test tasks that are untracked, just ignore the parameters.
- All other tasks set dirty bits to clean (0) on exit.
- Removed _DebugGroupTaskWrapper, UsdImagingGLEngine now just uses the task list from the task controller directly.

(Internal change: 1915068)