[Hd/Hdx] Move task's delegate getter helper to base class.
Continuing the goal of bringing the Task system more inline with the rest of the Sync algorithm,  this change replaces HdSceneTask::_GetSceneDelegateValue() with HdTask::_GetTaskParams().

This change is inline with the principle of Hd Prim classes providing a delegate wrapper function around delegate getters that access core parts of the schema for that prim type.

HdTask::_GetTaskParams() needs the delegate as a parameter as the base class doesn't store it (it is passed by Sync instead).  However, as the attribute id for the getter is fixed, _GetTaskParams() doesn't need GetSceneDelegateValue's valueId parameter

Changes:
- Added HdTask::_GetTaskParams().
- Removed HdSceneTask::_GetSceneDelegateValue()
- Substituted _GetTaskParams() for _GetSceneDelegateValue() in all callers.

(Internal change: 1915333)