[Hd/Hdx] Remove dependance on HdSceneTask for delegate and scene task.
In aiming to remove HdSceneTask, this change removes tasks's dependencies on HdSceneTask to provide the delegate and dirty bits.

HdSceneTask provided three methods for the GetDelegate(), _GetTaskDirtyBits() and _GetTaskDirtyState().

For the most part, the transition is simple as the delegate and  dirty bits are now passed as parameters to Sync.

_GetTaskDirtyState() returned both the dirty bits and the version number for the "geometry" collection.  This would have been incorrect behavior in regard to collections, but no tasks consumed this information anyway, so this wasn't an issue and could be safely removed.

GetDelegate() was a little trickier to remove as tasks used it outside of Sync.  For the most-part, it was just Sync helper functions, so these just needed to take the delegate as a parameter.

One more complex case was HdxDrawTargetTask.  In that task _Execute() was using the delegate to access the taskSetAlphaToCoverage enable flag.  Hydra doesn't expect that a delegate would change the return value from IsEnabled() for a given token over the life of a delegate, so the logic was moved to Sync using a member variable to pass the result to Execute.

Changes:
- Removed GetDelegate(), _GetTaskDirtyBits() and _GetTaskDirtyState() from HdSceneTask.
- In Tasks replaced _GetTaskDirtyBits() and _GetTaskDirtyState() by using the dirtyBits parameter passed to tasks.
- In Sync replaced GetDelegate() with parameter.
- Sync helper functions that used GetDelegate() now also receive delegate via a function parameter.
- In HdxDrawTargetTask, added _enableSampleAlphaToCoverage member attribute.
- Moved logic for if HdxDrawTargetTask should enable or disable sample alpha to coverage to Sync method.
- HdxDrawTargetTask execute method now uses _enableSampleAlphaToCoverage to determine if GL_SAMPLE_ALPHA_TO_COVERAGE should be enabled/disabled.

(Internal change: 1915373)