[Hd/Hdx] Remove HdSceneTask
The previous series of change lists have removed the need for HdSceneTask, so this change removes the class.
Ideally, with this change, the tasks that did depend on HdSceneTask no longer take the HdSceneDelegate as a parameter to the constructor.
To remove one of these dependencies in HdxRenderSetupTask, the creation of the render pass state is now deferred to first use.
Some other tasks have more complex dependencies, so this is deferred to later work.

Changes:
- Removed HdSceneTask.
- Switched tasks' base class to HdTask.
- Updated base class initialization in constructors.
- Added HdxRenderSetupTask::_GetRenderPassState() method to obtain the render pass state, so the defered creation is isolated to one place.
- Sync methods that used _renderPassState in HdxRenderSetupTask now access it through the new getter.

(Internal change: 1915415)