Fix inadvertent bump in required cmake version
Change 1915192 introduced the use of the DIRECTORY keyword
to the get_filename_component cmake command. This keyword
was introduced in cmake 2.8.12, which is later than the
documented requirement of 2.8.8 on Linux.

To avoid this bump we use the PATH keyword instead, which
is a legacy alias per the cmake docs.

(Internal change: 1915882)