tf: simplify implementation of TfWeakPtrFacade::PointsToA
The previous non-polymorphic overload failed to supply a return type
parameter to enable_if, making it effectively uncallable.  While we
could just remove that overload, also removing the enable_if and
static_assert from the other overload allows us to naturally cover
both cases.  gcc, clang and msvc all recognize the derived-to-base
dynamic_cast and generate only a null pointer check.  Similarly, the
static assertion is redundant as the compilers issue an error in
exactly the case it was covering.

(Internal change: 1922091)