[usdUtils] push GetMaterialsScopeName() from the Maya plugin down into usdUtils
Initially, the ability to configure the materials scope name was added to
the Maya plugin since that was a primary place where materials might be
authored. However, downstream USD consumers may also need to know the
"standard" scope name for a pipeline. This change moves the configurability
and querying of the materials scope name out of the Maya plugin and down
into the usdUtils library.

(Internal change: 1922841)