[usdImaging] Stop population traversal at non-imageable prims.
This change matches existing behavior in USD (for example, bbox cache traversal behavior).  Specifically, prune subtrees if they are (1) not imageable and (2) have a definite type (untyped scopes are ok).

Some additional related work:
- Remove IsPopulatedIndirectly(), as its functionality is replaced by the traversal check.
- Check that material relationships point to materials, before populating them by reference.
- Fix UsdImagingPrimAdapter virtual population flags: make them const-correct and un-inline them.
- Fix test assets with miscellanious misspellings of Xform.

This change also fixes a bug in point instancer population.  The population loop is essentially a for() over a subtree, but implemented as a while() with ++iter at the bottom for complicated reasons having to do with NI.  Previously, if you failed to populate a prim and continued, you'd end in an infinite loop since nobody would increment the loop pointer.  This change correctly increments the iterator when ShouldCullSubtree(*iter) is true and we continue; it also fixes up other cases of "continue" in the same function.

(Internal change: 1923540)
(Internal change: 1923672)