Deprecate UsdAttribute::HasAuthoredValueOpinion() in favor of HasAuthoredValue().  For historical reasons, the former returns true even if the attribute is blocked, which is not very useful.  Also beef up much related documentation.
We will retain HasAuthoredValueOpinion on UsdResolveInfo, for clients
that really want to know if there's *any* authored scene description
for an attribute, but it will eventually be removed for
UsdAttribute and UsdAttributeQuery.

With this change, we are now back to the same place we were before
we developed the blocking feature: HasValue() tells you whether an
attribute has any value source, and HasAuthoredValue tells you if
an attribute has a non-Fallback value source.

(Internal change: 1923544)