[Hd] When we reset varying state, only reset varying state for clean prims.
Hydra has a mechanism called "dirty lists" for caching what parts of the scene are likely to need scene updates each frame.  As prims are invalidated, they get the "varying" bit marked, and when hydra gathers a list of prims to sync it looks for that "varying bit".  When prims are cleaned (after syncing), the varying bit stays set so that the dirty list doesn't need to be rebuilt between frames.

The dirty lists need to be reset occasionally for this to provide an optimization; otherwise, the list of "dynamic" prims grows unbounded.  (Note that static geometry doesn't need to be included because the initial sync doesn't affect dirty list building).  The reset we've had for a long time is to wait until N% of the dirty list is clean, then reset the varying state for all rprims and rebuild dirty lists based on what's marked dirty the next frame.

I believe this is incorrect: for example, if renderpass A renders everything in the scene starting at prefix /A, and renderpass B renders prefix /B, imagine if renderpass A reset its dirty list while prims in the /B subtree were still dirty.  The prims in /B would have their varying bit removed, and the varying state would be incremented causing renderpass B to rebuild its dirty list.  However, it wouldn't pick up any of the prims since the dirty list build relies exclusively on the "Varying" bit.  I haven't definitively caught this case in the wild, but it could happen if you render e.g. a picking pass before the main render pass, since picking tends to use complicated collections.

Another issue is that a year and change ago we added an awesome and disruptive change to skip syncing invisible prims.  As an optimization, we kick them out of the dirty list, and don't let them back in until their visibility is marked dirty.  This saves a ton of pre-sync work on scenes where lots of things are invisible.  This is implemented by: if we skip sync on a prim, clear the varying bit; and don't add the varying bit back if you call MarkRprimDirty() with redundant dirty bits.  So if you call MarkRprimDirty(DirtyPoints) each frame, but the prim is invisible, it stays off the dirty list until you call MarkRprimDirty(DirtyVisibility) and force it to go through a full sync: if DirtyVisibility is set on a prim, we force it to go through sync, and setting it when its not set already will add the varying bit back.

The issue here is that if you have a visible prim which has lost its varying bit, for some reason, the early out in MarkRprimDirty means it will never be added to the dirty list no matter how much you invalidate it.  We want to keep this MarkRprimDirty optimization, since it helps us with large numbers of invisible prims; that means we need to make sure visible, non-clean prims never lose their varying bit.

That leads us to the change here: when we rebuild all the dirty lists, only remove the varying bit for clean prims, rather than for all prims.

The outcome is that dirty prims would stay on the dirty list, but either:
- (1) If they're included on a currently processed dirty list, and visible, they will be synced and cleaned.
- (2) If they're included on a dirty list for a renderpass we're not processing right now, they don't cause a slowdown for us, and they'll be handled correctly when that dirty list is synced.
- (3) If the prims are invisible the varying bit will be cleared by the sync-skipping functionality.

The main risk here is with Sync() calls that don't fully clear the dirty bits.  For example, it's possible for mesh to leave DirtySmoothNormals set if it's only drawing in flat shading that frame.  The previous code might reset the varying bit, pushing it off the dirty list; this would be more performant, but less correct since if you switched to smooth shading without Syncing that prim you'd render with stale smooth normal data.

(Internal change: 1924090)