[usdImaging] re-organize previewSurface.glslfx so that sections can be extracted and used by Maya
Maya's fragment graph system for Viewport 2.0 defines shading networks for
the viewport using XML files with embedded shader code. In developing the
Maya-equivalent of the UsdPreviewSurface shader, we want the Maya shader to
be using exactly the same code as much as possible, but forking the code by
copying and pasting excerpts from previewSurface.glslfx into the Maya
fragments could lead to a whole slew of maintainability and divergence
issues.

This change re-organizes the code in previewSurface.glslfx so that specific
sections ('Preview.LightStructures' and 'Preview.Lighting') of it can be
extracted and injected into the Maya fragment XML files programatically.
Function parameters were adjusted as well so that inputs from Maya can be
wired in directly.

Note also that the main entry point on the Hydra side "evaluateLights" was
broken out into its own "Preview.LightIntegration" section. This is
necessary because we want to make sure that no Hydra-specific code is
injected into the Maya fragments (e.g. lightSource data structures and
functions and other macros that come from Hydra's code generation).

(Internal change: 1924425)