Fix redundant node errors with specializes arcs
Composition would run into redundant node errors in cases
where a prim had both direct and ancestral specializes arcs.
This was because composition arcs that had already been
evaluated in the subgraph of the direct specializes nodes
were re-evaluated after that subgraph had been propagated
back to the origin of the ancestral specializes arc.

To avoid this, composition now propagates the shallowest
nodes beneath the direct specializes nodes to the origin.
Arcs beneath these propagated nodes will be evaluated as
normal and should yield the same subgraph as beneath the
direct specializes node. But, since that subgraph wasn't
propagated to the origin, we should avoid the redundant
node errors.

(Internal change: 1924790)