[pxrUsdPreviewSurface] introduce a UsdPreviewSurface-equivalent Maya shader node
This change adds a "pxrUsdPreviewSurface" Maya shader node intended to
provide the Maya-side equivalent of the UsdPreviewSurface shader in USD.
The Maya shader has all of the same attributes as UsdPreviewSurface, and in
fact uses much of the same code.

The shader executes in Viewport 2.0 using Maya's fragment graph system. The
fragment graph and the fragment XML files used by the surfacing shading
node override are generated by the pxrUsdPreviewSurfaceGenFragments.py
script, a typical invocation of which might look like this:

  ./pxrUsdPreviewSurfaceGenFragments.py \
      ../../../../pxr/usdImaging/lib/usdShaders/shaders/previewSurface.glslfx \
      --schema <Maya installation dir>/bin/ShadeFragment/v2pFragment.xsd

This script will lift the lighting/material-specific portions of shading
code from the provided path to the preview surface glslfx file. It will
also use the template XML and _shaderCode.glsl files in the source
directory to generate the final XML files for installation.

This generation script is also tied into the build system to run in a
validation mode when the plugin is built and ensure that the XML files are
up-to-date.

This change also adds USD export support for the new shader when using the
"useRegistry" shading mode.

(Internal change: 1925263)