Spruce up Primvars API's enhancing documentation, fixing behavior of inherited primvars for builtin attributes (e.g. primvars:displayColor on Gprims) and blocked attributes, encompassing complete inheritance behavior inside the API, rather than requiring clients to implement it themselves, because it is subtle and even we were getting it wrong, and improving error handling.
The logic in UsdGeomPrimvarsAPI() was using the existence of
*any* attribute definition on a prim to decide to prune the
ancestor-walk regardless of whether the primvar provides a value.  Instead,
we should be using UsdAttribute::HasAuthoredValue(), which takes both
blocks and fallbacks into account properly.  While
it is true that we need to read the attribute's value in order to
answer HasAuthoredValue() for blocked attributes, this should not result in
fetching any big data for usda or usdc files (sorry Alembic).

To help document the new semantics and make them consistent,
we also introduce UsdGeomPrimvar::HasAuthoredValue(), and
turn FindInheritedPrimvar into FindPrimvarWithInheritance() so that
clients do not need to resolve parent inheritance themselves.
Similarly, we change FindInheritedPrimvars() into FindInheritablePrimvars()
to highlight its purpose of efficiently computing inherited primvars
in a top-down tree traversal.

The handling of invalid prims in UsdGeomPrimvarsAPI was also
non-standard, being overly accommodating of what id truly an error.
For now, all such guards have been turned into coding errors, but
we would prefer to remove the tests altogether to achieve the same failure
mode as core methods.

FInally, take back inheritance-related API on UsdGeomImageable, which
is new and unused; and deprecate all other primvar-related API on Imageable,
which is subsumed by the more general UsdGeomPrimvarsAPI.

(Internal change: 1928592)
(Internal change: 1928766)
(Internal change: 1928933)