[usdMaya] source-only plugs with default values should return false for UsdMayaUtil::IsAuthored()
This change fixes the initial test of UsdMayaUtil::IsAuthored() to check
whether the plug is the destination of a connection instead of whether it
has *any* connection as an initial test for authored-ness. This resolves an
issue where plugs that have not been changed from their default value but
are the source (but *not* the destination) of a connection would be
considered authored.

(Internal change: 1928724)