Restore compatibility with Katana 2.5
The recently-added vtKatana library requires API added in
Katana 3.0, but the currently-stated requirement is 2.5v1.
To avoid having to bump up the required version for now,
this change adds version guards around the vtKatana library
and its client code so that it will be only be used with
3.0+. The code that existed previously has been restored
and will be used for older Katana versions.

(Internal change: 1929516)