[usdMaya/pxrUsdMayaGL] clean up pxrUsdProxyShape purpose-based draw toggle handling
This change adds the missing attribute for toggling drawing of "proxy"
purpose prims.

Note that the long names of the Maya attributes for toggling "render" and
"guide" purpose prims were changed from "displayRenderGuides" to
"drawRenderPurpose" and "displayGuides" to "drawGuidePurpose", respectively.
These names more accurately reflect what the attributes control. We expect
that most uses of pxrUsdProxyShape are transient from loading the
"Collapsed" representation of a pxrUsdReferenceAssembly, and/or that the
purpose-based draw toggle attributes are relatively rarely used, so the
name change should hopefully not cause much of an issue.

(Internal change: 1929517)