On the mac, HdStream does not yet work. If a user does not pass a --renderer opinion via command line, then we automatically fallback to use the legacy renderer. In this workflow, we have an issue since GetRendererDisplayName will receive an empty id, meaning "use HydraDisabled", but it thinks that Hydra is still enabled (since the environment variable is not set, because it is only set when --renderer is actually set) and we will try to get the plugin description causing a python error.
The fix is quite self-contained but here is an explanation of how the automatic renderer fallback works :

In python usdview _init, we have the ability to provide information to the user in case the user does not know what renderers are available in that machine. That happens via GetRendererOptionChoices, which calls to GetRendererPlugins (this function will check the HD_ENABLED flag, in case the user specificly requested to disable Hydra) and it returns a list of renderers. Then, we resolve their names via GetRenedrerDisplayName (which does not check with empty id, but it checks for HD_ENABLED disabled). On a mac, this function will end up returning HdStreamRendererPlugin and HydraDisabled.

The next important event happen during construction of the UsdImagingGL/engine where we check if Hydra is enabled in three ways :
   - Check if there is a gl context
   - Check the env variable HD_ENABLE which is set when the user passes the command line "--renderer HydraDisabled"
   - Try to get default plugin which will call to the plugin IsSupported, in the case of HdSt it will check if the context supports 4.0
On the mac, Hydra Stream is not supported, so we disable Hydra and construct a legacy renderer.

The construction of the engine class can be triggered by stageView, since we call glDraw which ends up calling getRenderer (if no renderer) and it calls GetRendererDisplayName to resolve the name with an empty id which causes the problem.

(Internal change: 1929729)