Fix bug when namespace editing a prim that's part of a chain of composition arcs.
The example case where this came up involved the following
authored composition arcs:

/Root/C -(specializes)-> /Root/B
/Root/B -(specializes)-> /Root/A

When /Root/A was renamed to /Root/A_New, Pcp recognized
that the specializes arc authored on /Root/B had to be
updated to point to the new path. This is the only update
needed in this case; /Root/C does not directly specialize /Root/A
so it doesn't need to be changed. However, Pcp failed to
recognize this, which ultimately led to the specializes arc on
/Root/C being cleared out entirely.

Pcp already had logic to determine that /Root/C did not need
to be updated in the _AddLayerStackSite function, but that
wasn't being invoked. It is now.

(Internal change: 1930005)