Another fix for build errors on Travis
Travis assumes that any command that does not produce
output for 10 minutes is frozen and errors out in
that case. Building USD itself runs into this issue
after the reduction in verbosity in change 1930451.

To avoid this, we use the travis_wait command when
running build_usd.py. Note that this means we no
longer get continuous log output until the script
finishes.

(Internal change: 1931451)