02 Add option to ImagingGL engine to render into its own internally manages 16F color and 32F depth targets.
This is needed for the color correction post process where we first need to render linear colors to 16F and then copy the result back to the clients (QglWidget) 8bit color buffers.

Also tried setting 16f buffers in QGlWidget (.setRedBufferSize) so we would not need to manage an internal framebuffer.
This failed to create 16f target and I believe that is because X11 does  not support 16f windows, so QT is unable to create it and fallsback to 8bit default framebuffer.

Depth needs to also be copied back to the client because e.g. UsdView will render additional things on top such as bounding boxes.

We use HdxCompositor to copy the color and depth back so that it doesn't matter what depth buffer the client has bound versus what depth buffer we use internally. (Trying to bind the same depth buffer as the client widget can be tricky because OpenGL misses API to query some information from the default framebuffer. Plus using the compositor makes the code change simpler & smaller)

(Internal change: 1932360)