Change SetDefaultValue on SdfPropertySpec so that it cannot load a plugin.
Since unloaded types still have TfTypes, we can compare still compare the TfTypes of the property and the attempted authored value to see if we should actually write to the layer. We duplicate the behavior if we know that the property's type is an enum, so for both unloaded and enum types (which should be one in the same) we only compare TfTypes. This means that authoring an integer value into the layer for an enum-typed attribute is no longer possible. All the cases that this happens have already been dealt with.

(Internal change: 1932403)